requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
openai==0.28.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
